ddd_particles doesn't overwrite _orig files

+++ programs/e2ddd_particles.py
@@ -121,6 +121,7 @@ def main():
 		elif os.path.exists("particles/{}_ptcls.hdf".format(base)) : src="particles/{}_ptcls.hdf".format(base)
 		dest="particles/{}__orig.hdf".format(base)
 		try: 
+			if os.path.exists(dest) : raise Exception
 			os.rename(src,dest)
 			if options.verbose>1: print "Renaming {} to {}".format(src,dest)
 			file(src,"w").write(file(dest,"r").read())			# copy the original data back to the source file so we don't have gaps for unaligned particles, but only if the rename worked
